JPEG 2000: per-tile COD/QCD overrides; 9/7 in single precision
Parse tile-part COD and QCD markers (not just RGN/POC) and resolve every coding
and quantization parameter per tile — p1_04 gives each tile its own QCD, which
previously made most tiles look over-coded and decline. All of buildTileComp,
the packet order, the tier-1 bit-plane count and the dequantization now use the
tile-resolved parameters.

Perform the irreversible 9/7 dequantization and inverse transform in float32 to
match the reference decoder's single-precision arithmetic (jpxFBand is now
[]float32). The reversible files stay bit-exact and p0_09 remains bit-exact.

p1_04 now decodes (was declining); it is a lossy 9/7 stream and reconstructs
within a fraction of a step of OpenJPEG on a small set of coarsely-quantized
coefficients.

+++ jpx_color.go
@@ -58,7 +58,7 @@ func decodeJPXImage(im *jpxImage) image.Image {
 				if b == nil {
 					return nil
 				}
-				placePlane(planes[c], im.xsiz, im.ysiz, b.x0, b.y0, b.w, b.h, comp.dx, comp.dy, func(i int) float64 { return b.data[i] })
+				placePlane(planes[c], im.xsiz, im.ysiz, b.x0, b.y0, b.w, b.h, comp.dx, comp.dy, func(i int) float64 { return float64(b.data[i]) })
 			}
 		}
 	}
